Aftermath changes 1

+++ database/migrations/2024_08_06_094630_create_seats_table.php
@@ -13,18 +13,15 @@ return new class extends Migration
     {
         Schema::create('seats', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('room_id')->default(1);
+            $table->integer('room_id')->default(1);
             $table->integer('row')->default(1);
             $table->integer('price')->default(100);
             $table->boolean('is_vip')->default(false)->nullable();
             $table->boolean('is_blocked')->default(false)->nullable(); 
             $table->timestamps();
-    
 
             $table->foreign('room_id')->references('id')->on('rooms')->onDelete('cascade'); 
-
         });
-    
     }
 
     /**
